requirements.txt
discord.py
python-dotenv
supabase
requests